Complete reconnection logic for wifi and 4G access
Disruption of both WiFi and the upstream (4G) internet connection are not catered for gracefully. Added ping test to establish if internet connection is live. A load of globals relating to the lemon telemetry have been removed and a lemon telemetry struct is used instead to store the latest gps and imu data. Also enabled checksum checks on uplink data from Mako. Removed redundant code.